Rename layers folders
