+++ napari/layers/base/_constants.py
