+++ napari/layers/base/_visual_wrapper.py
