+++ napari/layers/base/model.py
