+++ napari/layers/image/__init__.py
