+++ napari/layers/image/_constants.py
