+++ napari/layers/image/model.py
