+++ napari/layers/labels/_constants.py
