+++ napari/layers/labels/model.py
