+++ napari/layers/points/__init__.py
