+++ napari/layers/points/_constants.py
