+++ napari/layers/points/model.py
