+++ napari/layers/pyramid/__init__.py
