+++ napari/layers/shapes/__init__.py
