+++ napari/layers/shapes/_constants.py
