+++ napari/layers/shapes/mesh.py
