+++ napari/layers/shapes/model.py
