+++ napari/layers/shapes/shape_list.py
