+++ napari/layers/shapes/shape_util.py
