+++ napari/layers/shapes/shapes/ellipse.py
