+++ napari/layers/shapes/shapes/line.py
