+++ napari/layers/shapes/shapes/path.py
