+++ napari/layers/shapes/shapes/polygon.py
