+++ napari/layers/shapes/shapes/rectangle.py
