+++ napari/layers/shapes/shapes/shape.py
