+++ napari/layers/vectors/__init__.py
