+++ napari/layers/vectors/model.py
