+++ napari/layers/vectors/vectors_util.py
